# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"DefaultSelection_Prisma._36_FilesPayload_": {
				"properties": {
					"url": {
						"type": "string"
					},
					"fileType": {
						"type": "string"
					},
					"filesize": {
						"type": "number",
						"format": "double"
					},
					"filename": {
						"type": "string"
					},
					"fileId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"url",
					"fileType",
					"filesize",
					"filename",
					"fileId",
					"createdAt"
				],
				"type": "object"
			},
			"Files": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_FilesPayload_",
				"description": "Model Files"
			},
			"DefaultSelection_Prisma._36_UserPayload_": {
				"properties": {
					"avatarId": {
						"type": "number",
						"format": "double"
					},
					"githubLink": {
						"type": "string"
					},
					"facebookLink": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"password": {
						"type": "string"
					},
					"fullname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"avatarId",
					"githubLink",
					"facebookLink",
					"createdAt",
					"password",
					"fullname",
					"email",
					"username",
					"userId"
				],
				"type": "object"
			},
			"UserWithAvatarInterface": {
				"properties": {
					"avatarId": {
						"type": "number",
						"format": "double"
					},
					"githubLink": {
						"type": "string"
					},
					"facebookLink": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"password": {
						"type": "string"
					},
					"fullname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"avatar": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Files"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"avatarId",
					"githubLink",
					"facebookLink",
					"createdAt",
					"password",
					"fullname",
					"email",
					"username",
					"userId",
					"avatar"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DefaultSelection_Prisma._36_ProblemPayload_": {
				"properties": {
					"authorId": {
						"type": "number",
						"format": "double"
					},
					"memoryLimit": {
						"type": "number",
						"format": "double"
					},
					"timeLimit": {
						"type": "number",
						"format": "double"
					},
					"tags": {
						"type": "string"
					},
					"difficulty": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"problemId": {
						"type": "number",
						"format": "double"
					},
					"fileId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"authorId",
					"memoryLimit",
					"timeLimit",
					"tags",
					"difficulty",
					"status",
					"description",
					"title",
					"problemId",
					"fileId",
					"createdAt"
				],
				"type": "object"
			},
			"Problem": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_ProblemPayload_",
				"description": "Model Problem"
			},
			"SubmissionWithProblem": {
				"properties": {
					"stderr": {
						"type": "string"
					},
					"verdict": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"problemId": {
						"type": "number",
						"format": "double"
					},
					"submissionId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"problem": {
						"$ref": "#/components/schemas/Problem"
					}
				},
				"required": [
					"stderr",
					"verdict",
					"language",
					"code",
					"problemId",
					"submissionId",
					"createdAt",
					"userId",
					"problem"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAllACSubmissionsFromUserInterface": {
				"properties": {
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/SubmissionWithProblem"
						},
						"type": "array"
					}
				},
				"required": [
					"submissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseInterface_GetAllACSubmissionsFromUserInterface_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/GetAllACSubmissionsFromUserInterface"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateAvatarInterface": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/UserWithAvatarInterface"
					}
				},
				"required": [
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseInterface_UpdateAvatarInterface_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/UpdateAvatarInterface"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_UserPayload_",
				"description": "Model User"
			},
			"UserResponseInterface": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseInterface_UserResponseInterface_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/UserResponseInterface"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DefaultSelection_Prisma._36_ResultPayload_": {
				"properties": {
					"memory": {
						"type": "number",
						"format": "double"
					},
					"time": {
						"type": "number",
						"format": "double"
					},
					"testcaseIndex": {
						"type": "number",
						"format": "double"
					},
					"output": {
						"type": "string"
					},
					"resultId": {
						"type": "number",
						"format": "double"
					},
					"verdict": {
						"type": "string"
					},
					"submissionId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"memory",
					"time",
					"testcaseIndex",
					"output",
					"resultId",
					"verdict",
					"submissionId",
					"createdAt"
				],
				"type": "object"
			},
			"Result": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_ResultPayload_",
				"description": "Model Result"
			},
...
			"SubmitCodeResponseInterface": {
				"properties": {
					"submissionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"submissionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseInterface_SubmitCodeResponseInterface_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/SubmitCodeResponseInterface"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubmitCodeConfig": {
				"properties": {
					"code": {
						"type": "string"
					},
					"language": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"language"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProblemWithUserStatusInterface": {
				"properties": {
					"authorId": {
						"type": "number",
						"format": "double"
					},
					"memoryLimit": {
						"type": "number",
						"format": "double"
					},
					"timeLimit": {
						"type": "number",
						"format": "double"
					},
					"tags": {
						"type": "string"
					},
					"difficulty": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"problemId": {
						"type": "number",
						"format": "double"
					},
					"fileId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"userStatus": {
						"type": "boolean"
					}
				},
				"required": [
					"authorId",
					"memoryLimit",
					"timeLimit",
					"tags",
					"difficulty",
					"status",
					"description",
					"title",
					"problemId",
					"fileId",
					"createdAt",
					"userStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetOneProblemInterface": {
				"properties": {
					"problem": {
						"$ref": "#/components/schemas/ProblemWithUserStatusInterface"
					}
				},
				"required": [
					"problem"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseInterface_GetOneProblemInterface_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/GetOneProblemInterface"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubmissionWithResults": {
				"properties": {
					"stderr": {
						"type": "string"
					},
					"verdict": {
						"type": "string"
					},
					"language": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"problemId": {
						"type": "number",
						"format": "double"
					},
					"submissionId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"results": {
						"items": {
							"$ref": "#/components/schemas/Result"
						},
						"type": "array"
					}
				},
				"required": [
					"stderr",
					"verdict",
					"language",
					"code",
					"problemId",
					"submissionId",
					"createdAt",
					"userId",
					"results"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAllSubmissionsFromProblemInterface": {
				"properties": {
					"submissions": {
						"items": {
							"$ref": "#/components/schemas/SubmissionWithResults"
						},
						"type": "array"
					}
				},
				"required": [
					"submissions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseInterface_GetAllSubmissionsFromProblemInterface_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/GetAllSubmissionsFromProblemInterface"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TestcaseInterface": {
				"properties": {
					"input": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"output": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"input",
					"output"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetTestcasesInterface": {
				"properties": {
					"testcases": {
						"$ref": "#/components/schemas/TestcaseInterface"
					}
				},
				"required": [
					"testcases"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseInterface_GetTestcasesInterface_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/GetTestcasesInterface"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
...
		"/api/user/{userId}/submissions/AC": {
			"get": {
				"operationId": "GetACSubmissionsFromUser",
				"responses": {
					"200": {
						"description": "Successfully fetched submissions from user",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseInterface_GetAllACSubmissionsFromUserInterface_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/user/avatar": {
			"patch": {
				"operationId": "UploadAvatar",
				"responses": {
					"200": {
						"description": "Update avatar successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseInterface_UpdateAvatarInterface_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAvatar",
				"responses": {
					"200": {
						"description": "Delete avatar successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseInterface_UserResponseInterface_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/api/problems/{problem_id}": {
			"post": {
				"operationId": "Submit",
				"responses": {
					"200": {
						"description": "All testcases passed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseInterface_SubmitCodeResponseInterface_"
								}
							}
						}
					}
				},
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "problem_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SubmitCodeConfig"
							}
						}
					}
				}
			}
		},
...
		"/api/problems/no-account/{problem_id}": {
			"get": {
				"operationId": "GetOneProblemNoAccount",
				"responses": {
					"200": {
						"description": "Successfully fetched problem without user data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseInterface_GetOneProblemInterface_"
								}
							}
						}
					}
				},
				"description": "Fetch a single problem without user account data.",
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "problem_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/problems/with-account/{problem_id}": {
			"get": {
				"operationId": "GetOneProblemWithAccount",
				"responses": {
					"200": {
						"description": "Successfully fetched problem with user data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseInterface_GetOneProblemInterface_"
								}
							}
						}
					}
				},
				"description": "Fetch a single problem with user account data (status included).",
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "problem_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/problems/{problem_id}/submissions": {
			"get": {
				"operationId": "GetSubmissionsFromProblem",
				"responses": {
					"200": {
						"description": "Successfully fetched submissions from problem",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseInterface_GetAllSubmissionsFromProblemInterface_"
								}
							}
						}
					}
				},
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "problem_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/problems/{problem_id}/testcases": {
			"get": {
				"operationId": "GetTestcases",
				"responses": {
					"200": {
						"description": "Successfully fetched submissions from problem",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseInterface_GetTestcasesInterface_"
								}
							}
						}
					}
				},
				"tags": [
					"Problems"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "problem_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}